migrate useless users page
